requirements.txt

PyQt6==6.4.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
spacy==3.6.1
pdf2image==1.16.3
pytesseract==0.3.10
beautifulsoup4==4.12.2
python-docx==0.8.11
openpyxl==3.1.2
opencv-python-headless==4.8.0.76
optuna==3.3.0
joblib==1.3.2
Pillow==10.0.1